QCLI Automation for MCP Feature
